set more reasonable default driver query timeouts
ES is used in Tier 1 by the QGX, and so should have a relatively short
timeout (no one hop should take 3 minutes). The real time limiter should
be the cost of consecutive hops, which is governed by the job timeout.

Dgraph and Neo4j, OTOH, may be sent multihop queries with a disabled job
timeout. At that point, the limit should be permissive, and set mostly
to prevent queries clogging connections to the backend.