debug configs should set DEBUG
DEBUG makes only 1 worker run, as an async task instead of a new
process. This ensures that the worker doesn't get forcibly restarted
when it hangs on a breakpoint and such.

So, for the debugging config, we always want breakpoint support, so we
basically can't run multi-worker. So, only 2 configs which both set
DEBUG.