move entire internal logic to TypedDict instead of pydantic
Provides for greater consistency, at the cost of slightly more cursed
field access. Sorry future us, swapping back to dataclasses is gonna be
a bit painful.

This also fixes an issue with QGX failures due to mutation; we now use
the dict id instead of a hash to back-map to QEdgeIDs.