optimize operation table, fix prefixes
Previously the complexity was ~O(n) where n is the number of operations
in the table. Now it should be ~O(s + p + o) where it's the sum of
subject categories, predicates, and object categories in the search
critera, which is expected to always be smaller than n.

Additionally, metakg attributes are now advertized with correct
prefixes.